requirements.txt
aiofiles==24.1.0
aiogram==3.19.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
dotenv==0.9.9
environs==14.1.1
frozenlist==1.5.0
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
gspread==6.2.0
httplib2==0.22.0
idna==3.10
magic-filter==1.0.12
marshmallow==3.26.1
multidict==6.2.0
oauthlib==3.2.2
packaging==24.2
propcache==0.3.0
proto-plus==1.26.1
protobuf==6.30.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3
